feat(analytics): configure GA4 web stream
