+++ .env.example
@@ -1,3 +1,3 @@
-# Google Analytics Measurement ID (optional)
-# Get yours at https://analytics.google.com
+# Optional override for the production Google Analytics Measurement ID.
+# Useful when testing with a separate local or preview data stream.
 NEXT_PUBLIC_GA_ID=
