+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '@/styles/globals.css';
+import GoogleAnalytics from '@/components/GoogleAnalytics';
 import type { AppProps } from 'next/app';
 import { ThemeProvider } from 'next-themes';
-import Script from 'next/script';
 import dynamic from 'next/dynamic';
 
 const Analytics = dynamic(
@@ -12,26 +12,7 @@ const Analytics = dynamic(
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
-      {process.env.NEXT_PUBLIC_GA_ID && (
-        <>
-          <Script
-            strategy="afterInteractive"
-            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`}
-          />
-          <Script
-            id="google-analytics"
-            strategy="afterInteractive"
-            dangerouslySetInnerHTML={{
-              __html: `
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){dataLayer.push(arguments);}
-          gtag('js', new Date());
-
-          gtag('config', '${process.env.NEXT_PUBLIC_GA_ID}');`
-            }}
-          />
-        </>
-      )}
+      <GoogleAnalytics />
       <ThemeProvider
         enableSystem={true}
         attribute="class"
